Nearly all cases parsed with the new sbf parser
